+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change log for curry-base
 Under Development
 =================
 
+  * added type class extensions (merge branch `masterThesisBoehm`)
+
   * Removed record type extensions
 
   * Moved `CYT` monads to `curry-base` (`Curry.Base.Monad`)