+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Under Development
 
   * added type class extensions (merge branch `masterThesisBoehm`)
 
+  * Lexer is now capable of lexing binary integer literals, for instance
+    `0b101010` or `0B101010` can now be lexed and are converted to `42`.
+
   * Removed record type extensions
 
   * Moved `CYT` monads to `curry-base` (`Curry.Base.Monad`)